Rename embedded_can module.
